pyproject.toml
[project]
name = "kelp-aef"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "kyle-dorman", email = "kyle.e.dorman@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "geopandas>=1.1.1",
    "pyyaml>=6.0.3",
    "shapely>=2.1.2",
]

[project.scripts]
kelp-aef = "kelp_aef:main"

[build-system]
requires = ["uv_build>=0.11.3,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mypy>=2.0.0",
    "pre-commit>=4.6.0",
    "pytest>=9.0.3",
    "ruff>=0.15.12",
    "types-shapely>=2.1.0.20260408",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
mypy_path = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
